Set main module NAME to Gearman::Client since Gearman doesn't exist
MakeMaker's NAME argument has to be set to the "main module" of the
distribution and in this distribution it is considered Gearman::Client.

For comparison, Template-Toolkit has NAME set to 'Template' and
libwww-perl has NAME set to LWP.

This information doesn't do anything useful other
than making an entry in MYMETA files, perllocal.pod and (with cpanm)
install.json metadata.

I keep DISTNAME to Gearman so the tarball is still
Gearman-VERSION.tar.gz and that makes distribution information and
URLs on CPAN not change at all.